requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-MotorHAT==1.4.0
Adafruit-PureIO==0.2.3
aiohttp==3.4.4
async-timeout==3.0.1
asyncio-redis==0.15.1
attrs==18.2.0
chardet==3.0.4
flamegraph==0.1
idna==2.8
idna-ssl==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.5.2
numpy==1.15.4
prometheus-client==0.5.0
python-json-logger==0.1.11
PyYAML==3.13
redis==3.0.1
yarl==1.2.6
